Implements shadow jar and backfills README
This corrects shadow configuration and backfills documentation on how to
use it.

The shadow jar was not configured correctly, mostly due to the "spark"
scope, which was not necessary to call out. This removes that scope and
defers to "compile". The impact is that if a user expects to exclude
spark, or handle it differently, they should just exclude the transitive
dep.

This also renames the env variable `ZIPKIN_KEYSPACE` to
`CASSANDRA_KEYSPACE`. This properly aligns the namespace to upstream and
other cassandra-specific properties.