Ports the project to Java, using Retrolambda for JRE 7 + readability
This ports the scala code to Java, and adds the usual tooling, like
license headers, error-prone and animal sniffer checks, etc.

What's different is that this uses Retrolambda, a tool that rewrites
java 8 bytecode to java 7. I chose to use this as spark jobs are easier
to understand when written with lambdas.